backend/requirements.txt
click==7.1.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
psycopg2-binary==2.8.4
PyYAML==5.3
six==1.14.0
SQLAlchemy==1.3.15
Werkzeug==1.0.0

